// features/marketing/types/validation.ts
export const bannerSchema = z.object({
    title: z.string().min(1, "Title is required"),
    link: z.string().url("Invalid URL"),
    type: z.enum(['DESKTOP', 'MOBILE', 'TABLET', 'HOMEPAGE', 'CATEGORY', 'POPUP']),
    priority: z.number().int(),
    status: z.enum(['ACTIVE', 'INACTIVE']),
    image: z.any().refine((file) => file, "Image is required."),
});